+++ crates/executor/Cargo.toml
@@ -10,6 +10,7 @@ readme = "README.md"
 # reth 
 reth-primitives = { path = "../primitives" }
 reth-interfaces = { path = "../interfaces" }
+reth-revm = { path = "../revm" }
 reth-rlp = { path = "../rlp" }
 reth-db = { path = "../storage/db" }
 reth-provider = { path = "../storage/provider" }
