+++ crates/executor/src/lib.rs
@@ -7,12 +7,9 @@
 
 //! Reth executor executes transaction in block of data.
 
-pub mod config;
 pub mod eth_dao_fork;
 
 /// Execution result types
 pub mod execution_result;
 /// Executor
 pub mod executor;
-/// Wrapper around revm database and types
-pub mod revm_wrap;
