+++ crates/revm/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "reth-revm"
+version = "0.1.0"
+edition = "2021"
+license = "MIT OR Apache-2.0"
+repository = "https://github.com/paradigmxyz/reth"
+description = "reth specific revm utilities"
+
+[dependencies]
+# reth 
+reth-primitives = { path = "../primitives" }
+reth-interfaces = { path = "../interfaces" }
+reth-provider = { path = "../storage/provider" }
+
+revm = { version = "3.0.0"}
